verify() now checks the validity of each square, in addition to the validity of the rows and columns.  It also checks if each nonzero value in initialState is also in currentState.  I also spotted a logical error in the check for the rows and the columns.  That logical error has been corrected.
Some spelling and grammatical errors in the comments have also been
corrected.  The comments still use "nonzero" and "empty"
interchangeably, however.  I would like to change this ambiguity in the
future to make the comments more understandable.

Code has now been written for setCurrentState().

At this point, a test driver needs to be written to determine if each
method and constructor works properly.