My original idea was to use a variable of type byte to save on memory. However, due to the format for numbers stored in variables of type byte, I bit the bullet and changed the type for the array to int.
I have also creatd a test driver, Driver.java, to test all of the
methods in the Sudoku class.  This was very beneficial, as I found a bug
in the verify method in Sudoku.  All of the methods should work without
any problems now.

With the Sudoku class fully created, or at least created enough to be of
use to me, I will turn my attention to the problem of solving.  My
initial goal will be to implement several stochastic optimization
algorithms to solve the Sudoku.