I changed HillClimbing to make the constructor public.
I changed Driver to include a test case for HillClimbing.  I let
HillClimbing run for two minutes, but it didn't solve the example
sudoku.  I'll probably run the program again and walk away for about an
hour to see if it solves it, and if not, I'll have to tweak the
algorithm.